+++ common-utils/src/main/java/dev/avorakh/shop/common/utils/response/ErrorDetail.java
@@ -1,4 +1,4 @@
-package dev.avorakh.shop.function.model;
+package dev.avorakh.shop.common.utils.response;
 
 import lombok.*;
 import lombok.experimental.FieldDefaults;
