+++ get-products-by-id-function/build.gradle
@@ -12,6 +12,7 @@ dependencies {
     implementation project(":products-common")
     implementation project(":dynamodb-dao")
     implementation project(":product-svc")
+    implementation project(":common-utils")
 
     testImplementation project(":test-utils")
     testImplementation libs.bundles.groovy.test
