+++ get-products-by-id-function/src/main/java/dev/avorakh/shop/function/LambdaHandler.java
@@ -6,11 +6,11 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse;
 import com.amazonaws.services.lambda.runtime.logging.LogLevel;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import dev.avorakh.shop.common.utils.ResponseCommonUtils;
+import dev.avorakh.shop.common.utils.response.ErrorDetail;
+import dev.avorakh.shop.common.utils.response.ErrorResource;
 import dev.avorakh.shop.dao.DynamoDbProductDao;
 import dev.avorakh.shop.dao.DynamoDbStockDao;
-import dev.avorakh.shop.function.model.CommonUtils;
-import dev.avorakh.shop.function.model.ErrorDetail;
-import dev.avorakh.shop.function.model.ErrorResource;
 import dev.avorakh.shop.svc.DefaultProductService;
 import dev.avorakh.shop.svc.ProductService;
 import java.util.List;
@@ -54,23 +54,23 @@ public class LambdaHandler implements RequestHandler<APIGatewayV2HTTPEvent, APIG
                 var validationError = new ErrorDetail("the productId path parameter should be present");
                 var errorResource = new ErrorResource(VALIDATION_ERROR, 1001, List.of(validationError));
                 var body = objectMapper.writeValueAsString(errorResource);
-                return CommonUtils.toAPIGatewayV2HTTPResponse(400, body);
+                return ResponseCommonUtils.toAPIGatewayV2HTTPResponse(400, body);
             }
 
             var foundProduct = productService.get(id);
 
             if (foundProduct.isPresent()) {
                 var body = objectMapper.writeValueAsString(foundProduct.get());
-                return CommonUtils.toAPIGatewayV2HTTPResponse(200, body);
+                return ResponseCommonUtils.toAPIGatewayV2HTTPResponse(200, body);
             }
 
             logger.log("Product not found", LogLevel.ERROR);
 
-            return CommonUtils.toErrorAPIGatewayV2HTTPResponse(404, "PRODUCT_NOT_EXIST", 1100);
+            return ResponseCommonUtils.toErrorAPIGatewayV2HTTPResponse(404, "PRODUCT_NOT_EXIST", 1100);
         } catch (Exception e) {
             logger.log(e.getMessage(), LogLevel.ERROR);
 
-            return CommonUtils.toErrorAPIGatewayV2HTTPResponse(500, "INTERNAL_SERVER_ERROR", 1000);
+            return ResponseCommonUtils.toErrorAPIGatewayV2HTTPResponse(500, "INTERNAL_SERVER_ERROR", 1000);
         }
     }
 }
