+++ get-products-list-function/src/main/java/dev/avorakh/shop/function/LambdaHandler.java
@@ -6,9 +6,9 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse;
 import com.amazonaws.services.lambda.runtime.logging.LogLevel;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import dev.avorakh.shop.common.utils.ResponseCommonUtils;
 import dev.avorakh.shop.dao.DynamoDbProductDao;
 import dev.avorakh.shop.dao.DynamoDbStockDao;
-import dev.avorakh.shop.function.model.CommonUtils;
 import dev.avorakh.shop.svc.DefaultProductService;
 import dev.avorakh.shop.svc.ProductService;
 import lombok.AccessLevel;
@@ -43,10 +43,10 @@ public class LambdaHandler implements RequestHandler<APIGatewayV2HTTPEvent, APIG
 
             var body = objectMapper.writeValueAsString(productService.getAll());
 
-            return CommonUtils.toAPIGatewayV2HTTPResponse(200, body);
+            return ResponseCommonUtils.toAPIGatewayV2HTTPResponse(200, body);
         } catch (Exception e) {
             logger.log(e.getMessage(), LogLevel.ERROR);
-            return CommonUtils.toErrorAPIGatewayV2HTTPResponse(500, "INTERNAL_SERVER_ERROR", 1000);
+            return ResponseCommonUtils.toErrorAPIGatewayV2HTTPResponse(500, "INTERNAL_SERVER_ERROR", 1000);
         }
     }
 }
