+++ import-service/import-resource-file-function/build.gradle
@@ -8,7 +8,7 @@ dependencies {
     implementation platform(libs.aws.sdk.v2.bom.get().toString())
     implementation libs.bundles.aws.lambda
     implementation 'software.amazon.awssdk:s3'
-    implementation project(":products-common")
+    implementation project(":common-utils")
 
     testImplementation project(":test-utils")
     testImplementation libs.bundles.groovy.test
