+++ import-service/import-resource-file-function/src/test/groovy/dev/avorakh/shop/function/LambdaHandlerTest.groovy
@@ -3,7 +3,7 @@ package dev.avorakh.shop.function
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
-import dev.avorakh.shop.function.model.ErrorResource
+import dev.avorakh.shop.common.utils.response.ErrorResource
 import dev.avorakh.shop.localstack.LocalstackAwsSdkV2TestUtil
 import dev.avorakh.shop.function.test.TestContext
 import dev.avorakh.shop.function.test.TestUtils
