+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'java-aws-shop-api'
 
 // Common modules
+include 'common-utils'
 include 'products-common'
 include 'test-utils'
 include 'dynamodb-dao'
